project_name/requirements.txt
Django==1.10
django-registration==2.1.2
django-countries==4.0
django-debug-toolbar==1.6
django-material==0.12.5
django-cachalot==1.4.1
python-memcached==1.58
